refactor: 통계봇 - count 증가 / 저장 인터페이스 분리
