+++ blink-api/src/main/kotlin/com/jordyma/blink/infra/stats/LinkStatsIncreaseServiceImpl.kt
@@ -1,13 +1,14 @@
-package com.jordyma.blink.stats.service
+package com.jordyma.blink.infra.stats
 
 import com.jordyma.blink.redis.client.RedisClient
-import org.springframework.stereotype.Service
+import com.jordyma.blink.stats.service.LinkStatisticsService
+import com.jordyma.blink.stats.service.LinkStatsIncreaseService
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 
-@Service
-class LinkStatisticsService(private val redisClient: RedisClient) : LinkStatisticService {
-
+class LinkStatsIncreaseServiceImpl(
+    private val redisClient: RedisClient,
+) : LinkStatsIncreaseService {
     // INCR 한 뒤 '첫 생성'이면 TTL 부여
     private val incrExpireLua = """
         local val = redis.call('INCR', KEYS[1])
@@ -29,28 +30,17 @@ class LinkStatisticsService(private val redisClient: RedisClient) : LinkStatisti
     // TTL 3일
     private val ttlSeconds = (60 * 60 * 24 * 3).toString()
 
-
-    override fun getYesterdayLinkViewCount(): Int {
-        val yesterday = LocalDate.now().minusDays(1).format(DateTimeFormatter.ISO_DATE)
-        return redisClient.get("$LINK_VIEW_COUNT_KEY_PREFIX$yesterday")?.toIntOrNull() ?: 0
-    }
-
-    override fun getYesterdayDailyActiveUsers(): Int {
-        val yesterday = LocalDate.now().minusDays(1).format(DateTimeFormatter.ISO_DATE)
-        return redisClient.scard("$DAILY_ACTIVE_USERS_KEY_PREFIX$yesterday")?.toInt() ?: 0
-    }
-
     // 링크 조회 api 호출시 증가
     override fun incrementLinkView() {
         val today = LocalDate.now().format(DateTimeFormatter.ISO_DATE)
-        val key = "$LINK_VIEW_COUNT_KEY_PREFIX$today"
+        val key = "${LinkStatisticsService.LINK_VIEW_COUNT_KEY_PREFIX}$today"
         redisClient.eval(incrExpireLua, listOf(key), listOf(ttlSeconds))
     }
 
     // 일일 활성 사용자수 증가
     override fun recordUserActivity(userId: Long) {
         val today = LocalDate.now().format(DateTimeFormatter.ISO_DATE)
-        val key = "$DAILY_ACTIVE_USERS_KEY_PREFIX$today"
+        val key = "${LinkStatisticsService.DAILY_ACTIVE_USERS_KEY_PREFIX}$today"
         redisClient.eval(saddExpireLua, listOf(key), listOf(userId.toString(), ttlSeconds))
     }
